cleanup

+++ src/bdfi/bdfiAlg.java
@@ -4,7 +4,7 @@ public class bdfiAlg {
 
     //requires stars, count, currentReview >= 0
     public static int updateReview(int stars, int count, int currentReview) {
-        return Math.round((float) ((stars + count * currentReview) / ((float) (count + 1))));
+        return Math.round((stars + count * currentReview) / ((float) (count + 1)));
     }
 
 }
